end-to-end/requirements.txt
langchain==0.3.27
langchain-core==0.3.80
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-experimental==0.3.4
langchain-text-splitters==0.3.11
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.46

python-dotenv==1.2.1
wikipedia==1.4.0
arxiv==2.3.1
faiss-cpu==1.13.0
tiktoken==0.12.0

pydantic==2.12.4
pydantic-settings==2.12.0
ipykernel==7.1.0
numexpr==2.14.1
beautifulsoup4==4.14.2
requests==2.32.5
lxml==6.0.2

# Additional dependencies from requirements.prod.txt
aiohttp==3.13.2
dataclasses-json==0.6.7
openai==2.8.1
tenacity==9.1.2
jsonpatch==1.33
packaging==24.2
typing-extensions==4.15.0

# FastAPI and server dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.0.0
colorama==0.4.6
gunicorn==23.0.0

